backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
bcrypt==4.0.1
python-jose==3.3.0
passlib==1.7.4
groq==0.3.0
pillow==10.1.0
python-magic==0.4.27
aiofiles==23.2.1